requirements.txt
# Django and REST Framework
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-spectacular==0.27.2

# PostgreSQL adapter
psycopg2-binary==2.9.11

# Environment variables
python-decouple==3.8

# WSGI server for production
gunicorn==23.0.0

# Email functionality
django-email-verification==0.3.3
celery==5.4.0
redis==5.2.0

# Additional utilities
pytz==2024.2
